pyproject.toml
[tool.poetry]
name = "images"
version = "0.1.0"
description = ""
authors = ["Daniil Chibitock <daniil.chibitock@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
